Explode auth __init__ into packages
