+++ abe/templates/email_sign_in_sent.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta http-equiv="Content-type" content="text/html; charset=utf-8">
+    <meta name="viewport" content="width=device-width,initial-scale=1">
+    <meta name="description" content="Sign into ABE">
+    <title>ABE : Sign In</title>
+   </head>
+<body>
+    <h1>Sign into ABE</h1>
+    <div>
+        ABE has sent a link to your Olin email address. Click on this link
+        to sign into ABE.
+    </div>
+    <form action="{{ email_oauth_url }}" method="post">
+        {{ form.email.label }}: {{ form.email(placeholder='yourname@olin.edu') }}
+        {{ form.hidden_tag() }}
+        {{ form.submit() }}
+    </form>
+</body>
+</html>
