+++ tests/context.py
@@ -9,7 +9,7 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')
 
 MONGODB_TEST_DB_NAME = "abe-unittest"
 os.environ["DB_NAME"] = MONGODB_TEST_DB_NAME
-os.environ["INTRANET_IPS"] = "127.0.0.1/24"
+os.environ["INTRANET_CDIRS"] = "127.0.0.1/24"
 os.environ["MONGODB_URI"] = ""
 os.environ['SLACK_OAUTH_CLIENT_ID'] = 'slack-oauth-client-id'
